added next_tetraminos
Replaced the use of Tetraminotemplate in `bag` with just Tetramino (there is only one template, but many tetraminos)
Created a module for `Bag`, to simplify future merges, (I already edited logic.rs in another branch)
Created `next_tetraminos` because later it will be necessary to read and show the next tetraminos on the screen
